--- src/openaivec/aio/__init__.py
@@ -1,7 +1,0 @@
-from ..embeddings import AsyncBatchEmbeddings
-from ..responses import AsyncBatchResponses
-
-__all__ = [
-    "AsyncBatchEmbeddings",
-    "AsyncBatchResponses",
-]
